refactor: renmae routes
